# Full Stack Development/security-example/v1/helpers/swagger_api.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: Backend
  contact: {}
host: localhost:3000
basePath: /v1
securityDefinitions: {}
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /users:
    post:
      description: "Register a new user. Creates a user with the default `driver` and `admin` roles.\n\n`first_name` users first name\n\n`last_name` users surname\n\n`email` users contact email address. Must be unique.\n\n`username` the username with which user shall login. Must be unique.\n\n`password` the password to secure users account.\n\n`phoneNo` the users contact number and the number that will get the login OTT. Just add normal number here no need to add country prefix here as backend will do it for you. Just make sure to include the `countryCode` along with the request. There is a lookup API that checks if the number exists and add the relevant country prefix for you.\n\n`countryCode` needs to be ISO 3166-1 alpha-2. ([https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2))\n\n`roles` can be either `admin`, `customer` or `driver`.\n\nResponse:  \nReturns the created user record. `_id` is the user id."
      summary: Add user
      tags:
        - user
      operationId: Adduser
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/AdduserRequest'
      responses:
        '200':
          description: ''
          headers: {}
      security: []
    get:
      description: Get all users in the database.
      summary: Get users
      tags:
        - user
      operationId: Getusers
      deprecated: false
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      security:
        - bearer: []
...
definitions:
...
  AdduserRequest:
    title: AdduserRequest
    example:
      first_name: test
      last_name: test
      email: test4@gmail.com
      username: test4
      password: test4
      phoneNo: '000000000'
      countryCode: XX
      roles:
        - admin
        - driver
    type: object
    properties:
      first_name:
        type: string
      last_name:
        type: string
      email:
        type: string
      username:
        type: string
      password:
        type: string
      phoneNo:
        type: string
      countryCode:
        type: string
      roles:
        type: array
        items:
          type: string
    required:
      - first_name
      - last_name
      - email
      - username
      - password
      - phoneNo
      - countryCode
      - roles
...
security: []
...